requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
psycopg2-binary>=2.9.9
pydantic>=2.7.0
livekit>=1.1.2
livekit-agents>=1.4.2
livekit-plugins-openai>=0.10.0
livekit-plugins-sarvam>=0.3.0
livekit-plugins-silero>=0.7.0
livekit-plugins-elevenlabs>=0.7.0
livekit-plugins-deepgram>=0.7.0
openai>=1.30.0
python-dotenv>=1.0.0
pytz>=2024.1
aiohttp>=3.9.0
boto3>=1.34.0
python-json-logger>=2.0.7
tiktoken>=0.7.0
certifi>=2024.2.2
pgvector>=0.3.0
redis>=5.0.4
requests>=2.32.0
phonenumbers>=8.13.47
reportlab>=4.2.0
